+++ tests/__init__.py
@@ -98,12 +98,19 @@ METRICAL_POSITION_TESTFILES = [
     for fn in ["test_metrical_position.xml", "test_anacrusis.xml"]
 ]
 
+TIME_SIGNATURE_MAP_EDGECASES_TESTFILES = [
+    os.path.join(MUSICXML_PATH, fn)
+    for fn in ["test_ts_map_ts_starts_not_at_zero.xml"]
+]
+
 REST_ARRAY_TESTFILES = [
     os.path.join(MUSICXML_PATH, fn)
     for fn in ["test_unfold_complex.xml", "test_rest.musicxml"]
 ]
 
 NOTE_ARRAY_TESTFILES = [os.path.join(MUSICXML_PATH, fn) for fn in ["test_beats.xml"]]
+
+OCTAVE_SHIFT_TESTFILES = [os.path.join(MUSICXML_PATH, fn) for fn in ["example_octave_shift.musicxml"]]
 
 MERGE_PARTS_TESTFILES = [
     os.path.join(MUSICXML_PATH, fn)
@@ -172,6 +179,7 @@ KERN_TESTFILES = [
 
 KERN_TIES = [os.path.join(KERN_PATH, fn) for fn in ["tie_mismatch.krn"]]
 
+HARMONY_TESTFILES = [os.path.join(MUSICXML_PATH, fn) for fn in ["test_harmony.musicxml"]]
 
 MOZART_VARIATION_FILES = dict(
     musicxml=os.path.join(MUSICXML_PATH, "mozart_k265_var1.musicxml"),
@@ -194,4 +202,4 @@ WAV_TESTFILES = [
     ]
 ]
 
-PNG_TESTFILES = glob.glob(os.path.join(PNG_PATH, '*.png'))
+PNG_TESTFILES = glob.glob(os.path.join(PNG_PATH, "*.png"))